% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitgr.R
\name{obstools}
\alias{obstools}
\alias{fitgr}
\alias{smoothprev}
\title{Extract measurements from observations}
\usage{
fitgr(
  obsdata,
  winsize = 7,
  poscol = "pos",
  totcol = "tot",
  iscumul = TRUE,
  ksmooth = 7
)

smoothprev(
  obsdata,
  winsize = 7,
  poscol = "pos",
  totcol = "tot",
  iscumul = TRUE
)
}
\arguments{
\item{obsdata}{Data frame of observations.  Date in column date
daily normalized observations in column given in the \code{col} argument}

\item{winsize}{Window size calculating rates and positive fractions}

\item{poscol}{Name of the column holding the positive observations}

\item{totcol}{Name of the column holding the total observations}

\item{iscumul}{Flag indicating whether poscol and totcol are cumulative}

\item{ksmooth}{Smoothing length in days for growth rates.  Zero means no smoothing.}
}
\description{
Extract measurements of growth rates and sample prevalences from observed data.
The observations are relatively noisy, so windowed smoothing is used.
}
\details{
If \code{smooth} is set, the smoothing is done \emph{before} the growth rate
bias correction is applied.
}
\section{Functions}{
\itemize{
\item \code{fitgr}: Measure growth rate from observations

\item \code{smoothprev}: Measure sample prevalence with smoothing
}}

